Added tests for reader phase
Also fixed a bug when handling cyclic dependencies. A not implemented
error was raised, instead of a FatalPhaseError.

Now cycles are only detected when the methods of Dummy are accessed. This
seems to be the best place to raise cycle errors, since Dummy has been
introduced as cycle detector in the first place.