Implemented facilities to preserve bindings throughout rewriting
*This is a major rewrite*

As of #7, there have been problems with the binding of choices to
their corresponding dimensions. So I moved binding analysis to Namer phase.

As of #4, there have been problems with rewriting attributed trees. So
just using kiama's attribution system to store the bindings did not work.

There have been multiple threats:

(1) attributes are not preserved on the rewritten trees
(2) pointers to old binding instances are not updated, when the tree is
    modified.

Ad 1.
-----
Automatic copying the attributes on rewriting is now achieved using
a specialized subclass of kiama's `rewriting.Rewriter`. This
`SymbolPreservingRewriter` has to be mixed in using the cake pattern which
causes some issues in the tests (kiama's Rewriter introduces `equal` and
so does scalatest), but the advantages of later being able to abstract
over the concrete type of ASTNode outweigh.

Ad 2.
-----
To be able to update references to the binding instances of `Identifier`s
and `Choice`s the notion of `Symbol`s is introduced. Symbols are boxed
references and abstract over the concrete tree representation and the
underlying binding structure.

Rewriting DimensionChecker
--------------------------
In order to be able to use the new concept of symbols the dimension
checker had to be rewritten. Since we do now use direct references to
binding dimensions the notion of "free choices" within the dimension
checker became superfluous. The emerging graph structure is simpler
with the cost of loosing the "Dimension never used" warning.

Necessary Utils
---------------
The implementation of the symbol table required a hash map, based on
reference equality (IdentityHashMap).