Using DynamicVariable in the messaging system allows thread safe dynamic scope
Formerly dynamic scope of messaging properties such as `phase` or `level`
was implemented in a non thread safe way. Actually the implementation has
been pretty close to the one of `DynamicVariable` itself.

Switching to the thread safe DynamicVariable is thus a straightforward
change.

This also enables running the tests in parallel again